Started implementation of exception handling

+++ src/main/java/rs/raf/Airline/repositories/custom/implementation/CompanyServiceImp.java
@@ -2,6 +2,7 @@ package rs.raf.Airline.repositories.custom.implementation;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import rs.raf.Airline.exeptions.ApiRequestException;
 import rs.raf.Airline.model.Company;
 import rs.raf.Airline.model.Ticket;
 import rs.raf.Airline.model.dto.companyDto.AddCompanyDto;
@@ -36,6 +37,10 @@ public class CompanyServiceImp implements CompanyService {
     @Override
     public CompanyDto getCompanyAndTickets(Long id) {
         Company company = companyRepository.findById(id).get();
+
+        if(company == null)
+            throw new ApiRequestException("No company with that ID");
+
         List<Ticket> tickets = ticketRepository.findAllByCompany_Id(id);
         CompanyDto companyDto = new CompanyDto();
         companyDto.setName(company.getName());
@@ -45,12 +50,18 @@ public class CompanyServiceImp implements CompanyService {
 
     @Override
     public String deleteCompany(Long id) {
+        if(companyRepository.findById(id) == null)
+            throw new ApiRequestException("No company with that ID!");
         companyRepository.deleteById(id);
         return "Company Deleted!!!";
     }
 
     @Override
     public String addCompany(AddCompanyDto addCompanyDto) {
+        if(companyRepository.findCompanyByName(addCompanyDto.getName()) != null)
+            throw new ApiRequestException("Company with that name already exists!");
+        if(!addCompanyDto.getName().matches("[a-zA-Z0-9]*"))
+            throw new ApiRequestException("Company name must contain only letters and numbers");
         Company company = new Company();
         company.setName(addCompanyDto.getName());
         companyRepository.save(company);
@@ -59,9 +70,14 @@ public class CompanyServiceImp implements CompanyService {
 
     @Override
     public String editCompany(EditCompanyDto editCompanyDto) {
-        Company company = companyRepository.findById(editCompanyDto.getId()).get();
-        company.setName(editCompanyDto.getName());
-        companyRepository.save(company);
+        Company newCompany = companyRepository.findById(editCompanyDto.getId()).get();
+        Company company = companyRepository.findCompanyByName(editCompanyDto.getName());
+        if(company != null && !company.getId().equals(newCompany.getId()))
+            throw new ApiRequestException("Company with that name already exists!");
+        if(!editCompanyDto.getName().matches("[a-zA-Z0-9]*"))
+            throw new ApiRequestException("Company name must contain only letters and numbers");
+        newCompany.setName(editCompanyDto.getName());
+        companyRepository.save(newCompany);
         return "Company Edited!!!";
     }
